requirements.txt
cryptography
docker
requests
python-socketio
aiohttp
python-dotenv
pyjwt
redis
pika
marshmallow
